force float on nan remover
I got the error: 
TypeError: ufunc 'isnan' not supported for the input types, and the inputs could not be safely coerced to any supported types according to the casting rule ''safe''.
because the column happened to be of type str, even if it was a column of ints. matplotlib will pick this up can convert in its histogram anyway so it isn't an issue if the column is of type str. Alternatively, a try except could be made, so that this error is a little more intuitive.